uploaded all the .h and .cpp files that were created for this project along with the driver program
This project is a blood bank.In this project a hospital can manage all of the blood donors along with the type of blood each donor had.The hospital also shows how much the hospital will pay each donor including the travel expenses.This project demonstrates all the basic concepts of object oriented programming. I have also created a driver program that is a menu shown to the user to ask what action he wants to perfom.